refactor: move tests folder into src
