+++ io-ts/src/__tests__/Form.tsx
@@ -4,7 +4,7 @@ import user from '@testing-library/user-event';
 import { useForm } from 'react-hook-form';
 import * as t from 'io-ts';
 import * as tt from 'io-ts-types';
-import { ioTsResolver } from '../src';
+import { ioTsResolver } from '..';
 
 const schema = t.type({
   username: tt.withMessage(
